fix: to collect stats correctly, modify hnsw_storage_disk accepts algo_context_t as an parameter instead of thread_entry
