+++ src/storage/index_hnsw/hnsw_algo_common.hpp
@@ -1,0 +1,162 @@
+
+/*
+ *
+ * Copyright 2016 CUBRID Corporation
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+//
+// hnsw_algo_common.hpp
+//
+
+#ifndef _HNSW_ALGO_COMMON_HPP_
+#define _HNSW_ALGO_COMMON_HPP_
+
+#include <random>
+
+#include "hnsw_api.hpp"
+#include "hnsw_graph_base.hpp"
+#include "thread_entry.hpp"
+#include "vector_distance.hpp"
+
+namespace cubhnsw
+{
+  // =====================================================================
+  // algo's base structs
+  // =====================================================================
+  template <typename Traits>
+  struct candidate_t
+  {
+    using slot_id_t = typename Traits::slot_id_t;
+    distance_t distance;
+    slot_id_t slot;
+
+    candidate_t (distance_t distance, slot_id_t slot): distance (distance), slot (slot) {}
+    inline bool operator< (candidate_t other) const noexcept
+    {
+      return distance < other.distance;
+    }
+  };
+
+  template <typename Traits>
+  struct closer_candidate_t
+  {
+    bool operator() (candidate_t<Traits> const &a,
+		     candidate_t<Traits> const &b) const noexcept
+    {
+      return a.distance < b.distance; // min-heap or ascending
+    }
+  };
+
+
+  struct oid_hash
+  {
+    std::size_t operator() (const OID &o) const noexcept
+    {
+      std::size_t h = 0;
+      auto mix = [&h] (auto v)
+      {
+	std::size_t x = std::hash<std::decay_t<decltype (v)>> {} (v);
+	h ^= x + 0x9e3779b97f4a7c15ULL + (h << 6) + (h >> 2);
+      };
+
+      mix (o.volid);
+      mix (o.pageid);
+      mix (o.slotid);
+      return h;
+    }
+  };
+
+  struct oid_equal
+  {
+    bool operator() (const OID &a, const OID &b) const noexcept
+    {
+      return a.pageid == b.pageid
+	     && a.slotid == b.slotid
+	     && a.volid == b.volid;
+    }
+  };
+
+  template <typename T>
+  struct visit_set_helper
+  {
+    using type = std::unordered_set<T>;
+  };
+
+  template <>
+  struct visit_set_helper<OID>
+  {
+    using type = std::unordered_set<OID, oid_hash, oid_equal>;
+  };
+
+  template <typename Traits>
+  using visited_set_t = typename visit_set_helper<typename Traits::slot_id_t>::type;
+
+  template <typename Traits>
+  using candidates_view_t = std::vector<candidate_t<Traits>>;
+
+  template <typename Traits>
+  using candidates_allocator_t = std::allocator<candidate_t<Traits>>;
+
+  template <typename Traits>
+  using top_candidates_t =
+	  sorted_buffer_gt<candidate_t<Traits>, std::less<candidate_t<Traits>>, candidates_allocator_t<Traits>>;
+
+  template <typename Traits>
+  using next_candidates_t =
+	  max_heap_gt<candidate_t<Traits>, std::less<candidate_t<Traits>>, candidates_allocator_t<Traits>>;
+
+  template <typename Traits>
+  struct add_result_t
+  {
+    int error {NO_ERROR};
+    typename Traits::slot_id_t result;
+  };
+
+  template <typename Traits>
+  struct search_result_t
+  {
+    int error {NO_ERROR};
+    candidates_view_t<Traits> results {};
+    std::vector<OID> oids {};
+  };
+
+  template <typename Traits>
+  struct algo_context_t
+  {
+    top_candidates_t<Traits> m_top_candidates;
+    top_candidates_t<Traits> m_top_for_refine;
+    next_candidates_t<Traits> m_next_candidates;
+    visited_set_t<Traits> m_visits;
+    std::default_random_engine m_level_generator;
+    cubthread::entry *m_thread_p {nullptr};
+
+    // stats
+    bool m_is_perf_tracking {false};
+    std::size_t m_visited_nodes{};
+    std::size_t m_computed_distances{};
+    std::size_t m_computed_distances_in_refines{};
+    std::size_t m_computed_distances_in_reverse_refines{};
+
+    void clear_candidates ()
+    {
+      m_top_candidates.clear ();
+      m_next_candidates.clear();
+      m_visits.clear();
+    }
+  };
+}
+
+#endif // _HNSW_ALGO_COMMON_HPP_
